backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9


python-decouple==3.8
python-dotenv==1.0.1

scikit-learn>=1.4.0
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
django-filter==23.5
joblib==1.3.2
